==> share_ride/01_clean_and_merge_data.ipynb <==
#@title Import libraries

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns


from sklearn import linear_model
from sklearn.preprocessing import PolynomialFeatures
from sklearn.model_selection import train_test_split
from sklearn.model_selection import cross_validate
from sklearn.feature_selection import SelectFromModel
from sklearn.pipeline import Pipeline
from sklearn.model_selection import GridSearchCV
from sklearn.preprocessing import normalize
import missingno as msno

np.random.seed(42)
%matplotlib inline

df4 = pd.read_csv("April.csv")

df4.columns

df4=df4.drop(['Trip ID',  'Pickup Census Tract', 'Dropoff Census Tract', 'Shared Trip Authorized','Trips Pooled', 'Pickup Centroid Location',  'Dropoff Centroid Location'],axis=1)

df5 = pd.read_csv("May.csv")
df5=df5.drop(['Trip ID',  'Pickup Census Tract', 'Dropoff Census Tract', 'Shared Trip Authorized','Trips Pooled', 'Pickup Centroid Location',  'Dropoff Centroid Location'],axis=1)

df6 = pd.read_csv("June.csv")
df6=df6.drop(['Trip ID',  'Pickup Census Tract', 'Dropoff Census Tract', 'Shared Trip Authorized','Trips Pooled', 'Pickup Centroid Location',  'Dropoff Centroid Location'],axis=1)

df7 = pd.read_csv("July.csv")
df7=df7.drop(['Trip ID',  'Pickup Census Tract', 'Dropoff Census Tract', 'Shared Trip Authorized','Trips Pooled', 'Pickup Centroid Location',  'Dropoff Centroid Location'],axis=1)

df8 = pd.read_csv("August.csv")
df8=df8.drop(['Trip ID',  'Pickup Census Tract', 'Dropoff Census Tract', 'Shared Trip Authorized','Trips Pooled', 'Pickup Centroid Location',  'Dropoff Centroid Location'],axis=1)

df=df4.append(df5).append(df6).append(df7).append(df8)

df.to_csv('share_ride_data.csv', index = False)
